Fix a few compil minor warnings and Coverity minor issues

+++ common/eda_base_frame.cpp
@@ -680,7 +680,7 @@ void EDA_BASE_FRAME::OnPreferences( wxCommandEvent& event )
     if( manager )
         static_cast<EDA_BASE_FRAME*>( manager )->InstallPreferences( &dlg, hotkeysPanel );
 
-    for( int i = 0; i < book->GetPageCount(); ++i )
+    for( size_t i = 0; i < book->GetPageCount(); ++i )
         book->GetPage( i )->Layout();
 
     if( dlg.ShowModal() == wxID_OK )
